requirements.txt
accelerate==1.10.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
filelock==3.19.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.1
idna==3.10
inquirerpy==0.3.4
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
oracledb==3.3.0
packaging==25.0
pfzy==0.3.4
prompt_toolkit==3.0.52
psutil==7.1.0
pycparser==2.23
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==80.9.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0+cpu
tqdm==4.67.1
transformers==4.56.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
